+++ CHANGELOG.md
@@ -18,7 +18,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   No more parenthesis needed.
 - `Object.prototype.$try()` will return `undefined` now instead of `null` if it
   can't find the property.
-- `document.$cookies.remove()` returns the value of the cookie.
+- `document.$cookies.remove()` & `document.$cookies.set()` return the value of
+  the cookie.
 
 ### Removed
 - Command line tests—use the browser for now.